requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.9
django-allauth==0.45.0
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.2
django-environ==0.8.1
django-extensions==3.1.5
idna==3.3
oauthlib==3.1.1
Pillow==8.4.0
psycopg2-binary==2.9.2
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
